Change getInt to getLong for family member history

+++ common/src/main/java/org/endeavourhealth/getFHIRRecordAPI/common/dal/JDBCDAL.java
@@ -1001,7 +1001,7 @@ public class JDBCDAL implements AutoCloseable {
                     .setStatus(resultSet.getString("status"))
                     .setName(resultSet.getString("name"))
                     .setCode(resultSet.getString("code"))
-            .setPractitionerId(resultSet.getInt("practitionerId"));
+            .setPractitionerId(resultSet.getLong("practitionerId"));
             familyMemberHistoryList.add(familyMemberHistory);
         }
         return familyMemberHistoryList;
